functions3 completed

+++ exercises/functions/functions2.cairo
@@ -1,12 +1,10 @@
 // functions2.cairo
 // Execute `starklings hint functions2` or use the `hint` watch subcommand for a hint.
 
-// I AM NOT DONE
-
 fn main() {
     call_me(3);
 }
 
-fn call_me(num: usize) {
+fn call_me(num: u64) {
     println!("num is {}", num);
 }
